loomit_server/requirements.txt
boto3==1.7.60
botocore==1.10.60
certifi==2018.1.18
chardet==3.0.4
Django==2.0.1
django-formtools==2.1
django-storages==1.6.6
docutils==0.14
idna==2.6
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
Pillow==5.2.0
psycopg2==2.7.3.2
python-dateutil==2.7.3
pytz==2017.3
requests==2.18.4
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
